# docs/protocol/http/openapi-v1.json
{
  "components": {
    "parameters": {
...
      "RequestId": {
        "in": "header",
        "name": "X-Request-ID",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "responses": {
      "Problem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "Stable typed failure.",
        "headers": {
          "X-Request-ID": {
            "description": "Matches the request and JSON body request_id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ConnectionListResponse": {
        "additionalProperties": false,
        "description": "Return all persisted connection profiles in repository order.",
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ConnectionProfile"
            },
            "title": "Connections",
            "type": "array"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "connections"
        ],
        "title": "ConnectionListResponse",
        "type": "object"
      },
      "ConnectionProfile": {
        "additionalProperties": false,
        "description": "已持久化并带有身份与审计时间的 SSH 连接配置。",
        "properties": {
          "auth_kind": {
            "enum": [
              "password",
              "private_key"
            ],
            "title": "Auth Kind",
            "type": "string"
          },
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "credential_id": {
            "format": "uuid",
            "title": "Credential Id",
            "type": "string"
          },
          "display_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "favorite": {
            "default": false,
            "title": "Favorite",
            "type": "boolean"
          },
          "group_name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Name"
          },
          "host": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Host",
            "type": "string"
          },
          "passphrase_credential_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passphrase Credential Id"
          },
          "port": {
            "default": 22,
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port",
            "type": "integer"
          },
          "proxy_jump_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Jump Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "username": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          },
          "version": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "display_name",
          "host",
          "username",
          "auth_kind",
          "credential_id",
          "connection_id",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "ConnectionProfile",
        "type": "object"
      },
      "ConnectionProfileInput": {
        "additionalProperties": false,
        "description": "创建或更新 SSH 连接配置时由调用方提供的字段。",
        "properties": {
          "auth_kind": {
            "enum": [
              "password",
              "private_key"
            ],
            "title": "Auth Kind",
            "type": "string"
          },
          "credential_id": {
            "format": "uuid",
            "title": "Credential Id",
            "type": "string"
          },
          "display_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "favorite": {
            "default": false,
            "title": "Favorite",
            "type": "boolean"
          },
          "group_name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Group Name"
          },
          "host": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Host",
            "type": "string"
          },
          "passphrase_credential_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Passphrase Credential Id"
          },
          "port": {
            "default": 22,
            "maximum": 65535,
            "minimum": 1,
            "title": "Port",
            "type": "integer"
          },
          "proxy_jump_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Proxy Jump Id"
          },
          "username": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "host",
          "username",
          "auth_kind",
          "credential_id"
        ],
        "title": "ConnectionProfileInput",
        "type": "object"
      },
      "ConnectionResponse": {
        "additionalProperties": false,
        "description": "Return one persisted connection profile.",
        "properties": {
          "connection": {
            "$ref": "#/components/schemas/ConnectionProfile"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "connection"
        ],
        "title": "ConnectionResponse",
        "type": "object"
      },
...
      "DeleteResponse": {
        "additionalProperties": false,
        "description": "Return the determined deletion outcome for an identity.",
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "deleted"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
...
      "HostKeyCandidate": {
        "additionalProperties": false,
        "description": "等待用户确认或替换的远端 SSH Host Key 候选值。",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "fingerprint_sha256": {
            "pattern": "^SHA256:[A-Za-z0-9+/=_-]+$",
            "title": "Fingerprint Sha256",
            "type": "string"
          },
          "host": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Host",
            "type": "string"
          },
          "key_algorithm": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Key Algorithm",
            "type": "string"
          },
          "port": {
            "maximum": 65535,
            "minimum": 1,
            "title": "Port",
            "type": "integer"
          },
          "public_key_openssh_b64": {
            "title": "Public Key Openssh B64",
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "host",
          "port",
          "key_algorithm",
          "fingerprint_sha256",
          "public_key_openssh_b64"
        ],
        "title": "HostKeyCandidate",
        "type": "object"
      },
      "HostKeyInspectionRequest": {
        "additionalProperties": false,
        "description": "Describe one direct or single-hop Host Key inspection.",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "jump": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SshJumpRequest"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "connection_id"
        ],
        "title": "HostKeyInspectionRequest",
        "type": "object"
      },
      "HostKeyRecord": {
        "additionalProperties": false,
        "description": "连接已确认的 Host Key 历史记录。",
        "properties": {
          "confirmed_at": {
            "format": "date-time",
            "title": "Confirmed At",
            "type": "string"
          },
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "fingerprint_sha256": {
            "title": "Fingerprint Sha256",
            "type": "string"
          },
          "host_key_id": {
            "format": "uuid",
            "title": "Host Key Id",
            "type": "string"
          },
          "key_algorithm": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Key Algorithm",
            "type": "string"
          },
          "public_key_openssh_b64": {
            "title": "Public Key Openssh B64",
            "type": "string"
          },
          "replaced_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaced At"
          },
          "status": {
            "enum": [
              "active",
              "replaced"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "host_key_id",
          "connection_id",
          "key_algorithm",
          "fingerprint_sha256",
          "public_key_openssh_b64",
          "status",
          "confirmed_at"
        ],
        "title": "HostKeyRecord",
        "type": "object"
      },
      "HostKeyReplacementRequest": {
        "additionalProperties": false,
        "description": "Bind a new candidate to the exact trusted fingerprint being replaced.",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "expected_old_fingerprint": {
            "title": "Expected Old Fingerprint",
            "type": "string"
          },
          "fingerprint_sha256": {
            "pattern": "^SHA256:[A-Za-z0-9+/=_-]+$",
            "title": "Fingerprint Sha256",
            "type": "string"
          },
          "host": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Host",
            "type": "string"
          },
          "key_algorithm": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Key Algorithm",
            "type": "string"
          },
          "port": {
            "maximum": 65535,
            "minimum": 1,
            "title": "Port",
            "type": "integer"
          },
          "public_key_openssh_b64": {
            "title": "Public Key Openssh B64",
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "host",
          "port",
          "key_algorithm",
          "fingerprint_sha256",
          "public_key_openssh_b64",
          "expected_old_fingerprint"
        ],
        "title": "HostKeyReplacementRequest",
        "type": "object"
      },
      "HostKeyResponse": {
        "additionalProperties": false,
        "description": "Return one persisted Host Key trust record.",
        "properties": {
          "host_key": {
            "$ref": "#/components/schemas/HostKeyRecord"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "host_key"
        ],
        "title": "HostKeyResponse",
        "type": "object"
      },
      "JsonValue": {},
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Expose one bounded machine-readable failure without raw internals.",
        "properties": {
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "error_code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "SshJumpRequest": {
        "additionalProperties": false,
        "description": "Freeze one ProxyJump profile version with its transient credential.",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "passphrase_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Passphrase B64"
          },
          "password_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Password B64"
          },
          "private_key_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Private Key B64"
          },
          "profile_version": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "Profile Version",
            "type": "integer"
          }
        },
        "required": [
          "connection_id",
          "profile_version"
        ],
        "title": "SshJumpRequest",
        "type": "object"
      },
...
      "SshStatusResponse": {
        "additionalProperties": false,
        "description": "Return one safe SSH lifecycle status.",
        "properties": {
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "SshStatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Loopback-only API owned by the Rust-managed packaged child. The Runtime WebSocket is specified separately.",
    "title": "Harness Shell Private Python Runtime API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/connections": {
      "get": {
        "description": "Return all persisted profiles through the shared application handler.",
        "operationId": "list_connections_v1_connections_get",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionListResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "List Connections"
      },
      "post": {
        "description": "Validate and persist one complete connection profile.",
        "operationId": "create_connection_v1_connections_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Create Connection"
      }
    },
    "/v1/connections/{connection_id}": {
      "delete": {
        "description": "Delete one profile only when its domain ownership gates allow it.",
        "operationId": "delete_connection_v1_connections__connection_id__delete",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Delete Connection"
      },
      "get": {
        "description": "Return one profile by opaque identity.",
        "operationId": "get_connection_v1_connections__connection_id__get",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Get Connection"
      },
      "patch": {
        "description": "Replace one complete connection profile while preserving its identity.",
        "operationId": "update_connection_v1_connections__connection_id__patch",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Update Connection"
      }
    },
...
    "/v1/host-key-confirmations": {
      "post": {
        "description": "Persist a user's explicit first-trust Host Key confirmation.",
        "operationId": "confirm_host_key_v1_host_key_confirmations_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostKeyCandidate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostKeyResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Confirm Host Key"
      }
    },
    "/v1/host-key-inspections": {
      "post": {
        "description": "Inspect a direct or single-hop Host Key without retaining credentials.",
        "operationId": "inspect_host_key_v1_host_key_inspections_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostKeyInspectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SshStatusResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Inspect Host Key"
      }
    },
    "/v1/host-key-replacements": {
      "post": {
        "description": "Replace a trusted Host Key with an explicit compare-and-swap fingerprint.",
        "operationId": "replace_host_key_v1_host_key_replacements_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostKeyReplacementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostKeyResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Replace Host Key"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "8765"
        }
      }
    }
  ]
}